analyzed the patch output to create a diff output
analyzed the patch output to create a diff output file for each project for later thrashing frequency calculation.  The purpose of this step is to remove the unnecessary information and transform all lines into a "canonical" form by removing tabs and spaces and converting everything to lower case, hence, making the diff file more clear and easier to analyze.